// MiLicenciaCentro/src/app/interfaces/pago/PinesOlimpia/Consulta/CuotaCea.ts
export interface CuotaCea {
    ansv: number;
    categoria: string;
    estado: number;
    fechaPago: Date | string;
    numeroCuota: number;
    numeroPin: string;
    sicov: number;
    valorBanco: number;
    valorCuotaCea: number;
    valorPin: number;
    nombres: string;
    apellidos: string;
    email: string;
    numeroTelefono: string;
    idTipoIdentificacion: number;
    numeroIdentificacion: string;
}